% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{counting}
\alias{counting}
\title{Counting peptides}
\usage{
counting(dt)
}
\arguments{
\item{dt}{A data table or data frame.}
}
\value{
Data table containing counts of peptides for each protein.
}
\description{
Count the occurences of relevant peptides for each protein.
}
\examples{
counting(data)
counting(sample)
}
