% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{my_DT}
\alias{my_DT}
\title{Datatable output}
\usage{
my_DT(x, ...)
}
\arguments{
\item{x}{Matrix or a data frame containing data.}

\item{...}{Optional arguments to \code{datatable}.}
}
\value{
HTML datatable widget displaying data.
}
\description{
Create a HTML datatable widget containing data with export buttons using library DT.
}
\examples{
my_DT(df)
my_DT(matrix, colnames = FALSE)
}
